+++ Changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Log the versions of our low level libraries.
 - Component `Dropdown`.
+### Fixed
+- CSS issue where hidden content was scrollable
 
 ## 4.28.3 - 2021-08-11
 ### Fixed